+++ src/Exceptions/MediaCannotBeDeleted.php
@@ -3,7 +3,6 @@
 namespace Spatie\MediaLibrary\Exceptions;
 
 use Exception;
-use Spatie\MediaLibrary\Models\Media;
 use Illuminate\Database\Eloquent\Model;
 
 class MediaCannotBeDeleted extends Exception
@@ -12,6 +11,6 @@ class MediaCannotBeDeleted extends Exception
     {
         $modelClass = get_class($model);
 
-        return new static("Media with id `{$mediaId}` cannot be deleted because it does not exist or does not belong to model {$modelClass} with id {$model->id}");
+        return new static("Media with id `{$mediaId}` cannot be deleted because it does not exist or does not belong to model {$modelClass} with id {$model->getKey()}");
     }
 }