+++ src/FileManipulator.php
@@ -3,6 +3,7 @@
 namespace Spatie\MediaLibrary;
 
 use Storage;
+use Illuminate\Support\Str;
 use Illuminate\Support\Facades\File;
 use Spatie\MediaLibrary\Models\Media;
 use Illuminate\Contracts\Bus\Dispatcher;
@@ -25,9 +26,9 @@ class FileManipulator
      *
      * @param \Spatie\MediaLibrary\Models\Media $media
      * @param array $only
-     * @param bool $onlyIfMissing
+     * @param bool $onlyMissing
      */
-    public function createDerivedFiles(Media $media, array $only = [], $onlyIfMissing = false)
+    public function createDerivedFiles(Media $media, array $only = [], $onlyMissing = false)
     {
         $profileCollection = ConversionCollection::createForMedia($media);
 
@@ -40,13 +41,13 @@ class FileManipulator
         $this->performConversions(
             $profileCollection->getNonQueuedConversions($media->collection_name),
             $media,
-            $onlyIfMissing
+            $onlyMissing
         );
 
         $queuedConversions = $profileCollection->getQueuedConversions($media->collection_name);
 
         if ($queuedConversions->isNotEmpty()) {
-            $this->dispatchQueuedConversions($media, $queuedConversions);
+            $this->dispatchQueuedConversions($media, $queuedConversions, $onlyMissing);
         }
     }
 
@@ -55,9 +56,9 @@ class FileManipulator
      *
      * @param \Spatie\MediaLibrary\Conversion\ConversionCollection $conversions
      * @param \Spatie\MediaLibrary\Models\Media $media
-     * @param bool $onlyIfMissing
+     * @param bool $onlyMissing
      */
-    public function performConversions(ConversionCollection $conversions, Media $media, $onlyIfMissing = false)
+    public function performConversions(ConversionCollection $conversions, Media $media, $onlyMissing = false)
     {
         if ($conversions->isEmpty()) {
             return;
@@ -73,11 +74,11 @@ class FileManipulator
 
         $copiedOriginalFile = app(Filesystem::class)->copyFromMediaLibrary(
             $media,
-            $temporaryDirectory->path(str_random(16).'.'.$media->extension)
+            $temporaryDirectory->path(Str::random(16).'.'.$media->extension)
         );
 
         $conversions
-            ->reject(function (Conversion $conversion) use ($onlyIfMissing, $media) {
+            ->reject(function (Conversion $conversion) use ($onlyMissing, $media) {
                 $relativePath = $media->getPath($conversion->getName());
 
                 $rootPath = config('filesystems.disks.'.$media->disk.'.root');
@@ -86,7 +87,7 @@ class FileManipulator
                     $relativePath = str_replace($rootPath, '', $relativePath);
                 }
 
-                return $onlyIfMissing && Storage::disk($media->disk)->exists($relativePath);
+                return $onlyMissing && Storage::disk($media->disk)->exists($relativePath);
             })
             ->each(function (Conversion $conversion) use ($media, $imageGenerator, $copiedOriginalFile) {
                 event(new ConversionWillStart($media, $conversion, $copiedOriginalFile));
@@ -125,7 +126,7 @@ class FileManipulator
             return $imageFile;
         }
 
-        $conversionTempFile = pathinfo($imageFile, PATHINFO_DIRNAME).'/'.str_random(16)
+        $conversionTempFile = pathinfo($imageFile, PATHINFO_DIRNAME).'/'.Str::random(16)
             .$conversion->getName()
             .'.'
             .$media->extension;
@@ -144,11 +145,11 @@ class FileManipulator
         return $conversionTempFile;
     }
 
-    protected function dispatchQueuedConversions(Media $media, ConversionCollection $queuedConversions)
+    protected function dispatchQueuedConversions(Media $media, ConversionCollection $queuedConversions, $onlyMissing = false)
     {
         $performConversionsJobClass = config('medialibrary.jobs.perform_conversions', PerformConversions::class);
 
-        $job = new $performConversionsJobClass($queuedConversions, $media);
+        $job = new $performConversionsJobClass($queuedConversions, $media, $onlyMissing);
 
         if ($customQueue = config('medialibrary.queue_name')) {
             $job->onQueue($customQueue);