+++ tests/Feature/HasMediaTrait/DeleteMediaTest.php
@@ -4,6 +4,9 @@ namespace Spatie\MediaLibrary\Tests\Feature\HasMediaTrait;
 
 use File;
 use Spatie\MediaLibrary\Tests\TestCase;
+use Spatie\MediaLibrary\MediaLibraryServiceProvider;
+use Spatie\MediaLibrary\Tests\Support\TestModels\TestModel;
+use Spatie\MediaLibrary\Tests\Support\TestModels\TestCustomMediaModel;
 
 class DeleteMediaTest extends TestCase
 {
@@ -11,17 +14,7 @@ class DeleteMediaTest extends TestCase
     {
         parent::setUp();
 
-        foreach (range(1, 3) as $index) {
-            $this->testModel
-                ->addMedia($this->getTestJpg())
-                ->preservingOriginal()
-                ->toMediaCollection();
-
-            $this->testModel
-                ->addMedia($this->getTestJpg())
-                ->preservingOriginal()
-                ->toMediaCollection('images');
-        }
+        $this->addMedia($this->testModel);
     }
 
     /** @test */
@@ -36,9 +29,7 @@ class DeleteMediaTest extends TestCase
         $this->assertCount(0, $this->testModel->getMedia('images'));
     }
 
-    /**
-     * @test
-     */
+    /** @test */
     public function it_will_remove_the_files_when_clearing_a_collection()
     {
         $ids = $this->testModel->getMedia('images')->pluck('id');
@@ -54,9 +45,7 @@ class DeleteMediaTest extends TestCase
         });
     }
 
-    /**
-     * @test
-     */
+    /** @test */
     public function it_will_remove_the_files_when_deleting_a_subject()
     {
         $ids = $this->testModel->getMedia('images')->pluck('id');
@@ -66,6 +55,32 @@ class DeleteMediaTest extends TestCase
         });
 
         $this->testModel->delete();
+
+        $ids->map(function ($id) {
+            $this->assertFalse(File::isDirectory($this->getMediaDirectory($id)));
+        });
+    }
+
+    /** @test */
+    public function it_will_remove_the_files_when_using_a_custom_model_and_deleting_it()
+    {
+        config()->set('medialibrary.media_model', TestCustomMediaModel::class);
+
+        (new MediaLibraryServiceProvider($this->app))->boot();
+
+        $testModel = TestModel::create(['name' => 'test']);
+
+        $this->addMedia($testModel);
+
+        $this->assertInstanceOf(TestCustomMediaModel::class, $testModel->getFirstMedia());
+
+        $ids = $testModel->getMedia('images')->pluck('id');
+
+        $ids->map(function ($id) {
+            $this->assertTrue(File::isDirectory($this->getMediaDirectory($id)));
+        });
+
+        $testModel->delete();
 
         $ids->map(function ($id) {
             $this->assertFalse(File::isDirectory($this->getMediaDirectory($id)));
@@ -86,5 +101,20 @@ class DeleteMediaTest extends TestCase
         $ids->map(function ($id) {
             $this->assertTrue(File::isDirectory($this->getMediaDirectory($id)));
         });
+    }
+
+    private function addMedia(TestModel $model)
+    {
+        foreach (range(1, 3) as $index) {
+            $model
+                ->addMedia($this->getTestJpg())
+                ->preservingOriginal()
+                ->toMediaCollection();
+
+            $model
+                ->addMedia($this->getTestJpg())
+                ->preservingOriginal()
+                ->toMediaCollection('images');
+        }
     }
 }